fix selectBox module to organizationSelectBox module
